+++ o3research/marketing/__init__.py
@@ -7,6 +7,8 @@ from .budgeter import Budgeter
 from .funnel_planner import FunnelPlannerAgent
 from .creative_prompt_agent import CreativePromptAgent
 from .creative import CreativeAgent
+from .landing_page_agent import LandingPageAgent
+from .lead_capture_agent import LeadCaptureAgent
 from .optimizer import PerformanceOptimizer
 from .api_writer import push_campaign
 
@@ -18,6 +20,8 @@ __all__ = [
     "FunnelPlannerAgent",
     "CreativePromptAgent",
     "CreativeAgent",
+    "LandingPageAgent",
+    "LeadCaptureAgent",
     "PerformanceOptimizer",
     "push_campaign",
 ]